+++ components/Navigation.tsx
@@ -1,0 +1,7 @@
+import {Box} from "@mui/system";
+
+export default async function Navigation(){
+    return (
+        <Box>Navbar placeholder</Box>
+    )
+}
