+++ components/pages/Browse.tsx
@@ -1,0 +1,9 @@
+import Box from "@mui/material/Box";
+
+export default async function Browse(){
+    return (
+        <Box>
+           <p>Hi</p>
+        </Box>
+    )
+}
